+++ KiMaLi/app/src/main/java/com/example/kimali/compensation/compensation_firstActivity.kt
@@ -20,7 +20,6 @@ class compensation_firstActivity : AppCompatActivity() {
     private lateinit var mDatabase: DatabaseReference
     lateinit var total_money :String
     lateinit var total_pcTime :String
-    var client: MqttAndroidClient? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -35,52 +34,6 @@ class compensation_firstActivity : AppCompatActivity() {
         setTitle(name)
         mDatabase = FirebaseDatabase.getInstance().reference
 
-        val clientId: String = MqttClient.generateClientId()
-        Log.i("hyolls","자식 토픽값 : "+topic)
-        client = MqttAndroidClient(applicationContext, "tcp://broker.hivemq.com:1883", clientId)
-
-
-        //connect하는 부분
-        try {
-            Log.i("hyolls","확인")
-            val token = client!!.connect(getMqttConnectionOption())
-            token.actionCallback = object : IMqttActionListener {
-                override fun onSuccess(asyncActionToken: IMqttToken) { //연결에 성공한 경우
-                    Log.i("hyoriTopic", "connection1")
-                    try {
-                        client!!.subscribe("topic", 0) //topic값 받음
-                    } catch (e: MqttException) {
-                        e.printStackTrace()
-                    }
-                }
-
-                override fun onFailure(
-                    asyncActionToken: IMqttToken,
-                    exception: Throwable
-                ) { //연결에 실패한 경우
-                    Toast.makeText(applicationContext, "연결에 실패하였습니다...(1)", Toast.LENGTH_SHORT).show()
-                }
-            }
-        } catch (e: MqttException) {
-            e.printStackTrace()
-        }
-
-        client!!.setCallback(object : MqttCallback {
-            //콜백처리하는 부분
-            override fun connectionLost(throwable: Throwable) {
-                Toast.makeText(applicationContext, "연결이 끊겼습니다...", Toast.LENGTH_SHORT).show()
-            }
-
-            @Throws(Exception::class)
-            override fun messageArrived(topic: String, message: MqttMessage) {
-                if (topic == topic) { //topic별로 나누어서
-                    val msg = String(message.payload)
-                    Toast.makeText(applicationContext, msg, Toast.LENGTH_SHORT).show()
-                }
-            }
-
-            override fun deliveryComplete(iMqttDeliveryToken: IMqttDeliveryToken) {}
-        })
 
 
         mDatabase.child("mission").child(topic).addListenerForSingleValueEvent(
